+++ src/composables/dialog-state-composable.ts
@@ -2,7 +2,6 @@
 import { computed, ref } from 'vue';
 
 const uploadPhotoDialogState = ref(false);
-const photoDetailDialogState = ref(false);
 const updateAlbumDialogState = ref(false);
 const updateAlbumTagsDialogState = ref(false);
 const albumToBeUpdate = ref({ id: '', albumName: '', desc: '', tags: [], private: false } as Album);
@@ -10,10 +9,6 @@ const albumToBeUpdate = ref({ id: '', albumName: '', desc: '', tags: [], private
 export default function DialogStateComposable() {
   const setUploadPhotoDialogState = (state: boolean) => {
     uploadPhotoDialogState.value = state;
-  };
-
-  const setPhotoDetailDialogState = (state: boolean) => {
-    photoDetailDialogState.value = state;
   };
 
   const setAlbumToBeUpdated = (album: Album) => {
@@ -39,7 +34,5 @@ export default function DialogStateComposable() {
     getUpdateAlbumTagsDialogState: computed(() => updateAlbumTagsDialogState.value),
     setUpdateAlbumTagsDialogState,
     updateAlbumTagsDialogState,
-    getPhotoDetailDialogState: computed(() => photoDetailDialogState.value),
-    setPhotoDetailDialogState,
   };
 }
